Use pre_exec instead of deprecated before_exec
The std::os::unix::process::CommandExt implementation of before_exec is
falsely not marked `safe`, as it is executed in the fork, in a very
constrained environment.
Instead use the unsafe `pre_exec` method, which calls
`RunningJail::attach`. In the happy path, `attach` does not allocate,
and in the error path, it panics, so safety guarantees should be met.